Move code from main.go to core.go
